Licence compliance checker
This is a command line tool to check project dependencies do not
contain any restricted dependencies.

It uses `go-license-detector` to identify the licences for each project
and then validate that they do not match any of the restricted licences.

The list of restricted policies are passed as arguments to allow
projects to have a different restriction policies.

Optional filters can be used to ignore or override licences for one or
more projects