Use go-license-detector API rather than executable
This is so the library we publish is completely self-contained and does
not rely on `go-license-detector` being installed in the GOPATH.
This fix is using a forked version of the `go-licence-detector`
that contains an improvement to make the detection results part of the
API: src-d/go-license-detector#42. Once the PR is merged we will reuse main repo.

Also adds a CHANGELOG.md

Fix #2