Commit vendor subdirectory to source control
This is required in order to have an easy integration with `go get`, as
it has no integration with `dep`.

We still use `dep` for defining our dependencies and should use `dep
ensure` as normal to download dependencies.

Also corrected a typo in the README around use of `go get` to install
the tool.