Move main package into project root (#14)
For some reason, with the main package in the `cmd` subdirectory, `go get` and `go install` weren't building an executable for this licence checker. Since these are the two standard ways of fetching and
installing binaries, this was something of a shortcoming.

I've fixed it by moving the main.go and main_test.go into the project root. This also has the effect of slightly simplifying the `go get` command given in the README.